+++ edsnlp/pipes/trainable/relation_detector_ffn/factory.py
@@ -1,0 +1,9 @@
+from edsnlp import registry
+
+from .relation_detector_ffn import RelationDetectorFFN
+
+create_component = registry.factory.register(
+    "eds.relation_detector_ffn",
+    assigns=[],
+    deprecated=[],
+)(RelationDetectorFFN)